pyproject.toml
[project]
name = "blog-backend"
version = "0.1.0"
description = "Backend for Blog"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "mypy>=1.19.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "python-slugify>=8.0.4",
    "ruff>=0.14.10",
    "uvicorn>=0.40.0",
]

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
